# api/openapi.yaml
openapi: 3.0.3
info:
  title: Claims System API
  version: 0.5.0
  description: |
    Backend API for claims ingestion, reconciliation, policy administration, and reporting.

    **Sensitive fields:** When `API_REDACT_SENSITIVE_FIELDS` is true (default), claim JSON masks pharmacy NPI and redacts `hashed_rx_key`.
    CSV exports honor `EXPORT_REDACT_SENSITIVE_FIELDS` (default true) for the same columns.

    **Password reset:** Production sends the token by email (SMTP env vars). `PASSWORD_RESET_EXPOSE_TOKEN` returns the token in JSON for dev/tests only.
    `PASSWORD_RESET_PUBLIC_BASE_URL` is used in the email body for the API base URL (default `http://localhost:$PORT`).

    **Admin MFA:** When `REQUIRE_MFA_FOR_ADMIN=true`, users with an admin role cannot log in until MFA is enabled (`code` MFA_SETUP_REQUIRED).

    **MFA step-up:** When `REQUIRE_MFA_STEP_UP_FOR_ADMIN=true`, MFA-enabled admins must call `POST /v1/mfa/step-up` and send the returned token on the `X-MFA-Step-Up` header for mutating admin routes (organizations, manufacturers, policies, rebate records, contract pharmacies, dispute resolution).

    **Platform (v1):** Tenant admin routes under `/v1/organizations/{id}/` for settings, users, feature flags, SSO (OIDC + SAML), webhooks, and exception cases. See `CHANGELOG.md` and `docs/integrator-guide.md`.

...

    **SCIM 2.0:** `/scim/v2` with `Authorization: Bearer` (`SCIM_BEARER_TOKEN`); uses `SCIM_ORG_ID`.

...

...
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /v1/auth/saml/acs:
    post:
      summary: SAML Assertion Consumer Service (returns JWT + user)
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [SAMLResponse, RelayState]
              properties:
                SAMLResponse: {type: string}
                RelayState: {type: string}
      responses:
        "200": {description: Session token issued}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/UnauthorizedError"}
  /v1/auth/saml/metadata:
    get:
      summary: SAML SP metadata (register with IdP)
      security: []
      responses:
        "200":
          description: SAML metadata XML
          content:
            application/samlmetadata+xml:
              schema: {type: string}
        "404": {$ref: "#/components/responses/NotFoundError"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    scimBearer:
      type: http
      scheme: bearer
      description: Value must match SCIM_BEARER_TOKEN; SCIM_ORG_ID scopes the tenant.
...
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ValidationErrorResponse"}
    UnauthorizedError:
      description: Missing or invalid authentication credentials.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...
    NotFoundError:
      description: Requested resource does not exist.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error: {type: string}
        code: {type: string}
        request_id: {type: string}
        trace_id: {type: string, description: Correlates with X-Trace-Id and OpenTelemetry trace}
    ValidationIssue:
      type: object
      required: [field, message]
      properties:
        field: {type: string}
        message: {type: string}
    ValidationErrorResponse:
      type: object
      required: [error, code]
      properties:
        error: {type: string}
        code: {type: string, example: VALIDATION_ERROR}
        request_id: {type: string}
        details:
          type: array
          items: {$ref: "#/components/schemas/ValidationIssue"}
...